requirements.txt
# Core ML pipeline
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.44.0
Pillow>=10.0.0

# Detection & Segmentation
# OWLv2 via transformers (included above)
# SAM2 via transformers (included above)

# Identification
easyocr>=1.7.0
open-clip-torch>=2.24.0

# BGG matching
sentence-transformers>=2.2.0
requests>=2.31.0

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0
shapely>=2.0.0
